RTL/LTR rendering engine hinting
This prevents UI glitches with RTL feeds.

Use the unicode LTR marker character to hint the rending engine which
part of a RTL/LTR mixed string is LTR. This prevents UI glitches with
RTL feeds like http://www.ynet.co.il/Integration/StoryRss544.xml.

The entity has to be outside of the span tag to prevent flickering when
the counter is update via javascript.