introduce release script
This script replaces make-archive.sh and make-tag.sh.

Switch to 'git archive' to create a ZIP file from the repo. Its way
more faster, automates a lot things and works with the GitHub download
HEAD/Release/Tag as ZIP functionality as well.

git archive reads the file .gitattributes to determine which files has
to be excluded from the zip file.

It replaces the following placeholders in common.php:

$Format:%d$": with the output of git log --pretty=format:%d which
contains the tag of the last commit (if tagged).

$Format:%H$": with the hash of the latest included commit.

If no tag can be found the hash is appended to "master" and shown as
version number.

The approach is kind of hackish since no distinct format option for the
tag exists.

Fixes #231, https://github.com/fguillot/miniflux/issues/213#issuecomment-65631765