re-add css property appearance
I've removed the property a long time ago from the css with the hint
that it doesn't change any visible. That was wrong.

The property is related to mobile Safari. The appearance property is
needed to remove the rounded corners from buttons and to add the glow
effect to active input fields.