move the item action links to the top
there must be no space between between the <p class="preview"> and the $item['content']. The CSS pseudo-class :empty does not consider a paragraph with whitespaces as empty.  The Selectors Level 4 Draft[1] as of 13 September 2015, mentions the CSS pseudo-class :blank which will do so. But it's a draft and not yet implemented by the majority of the browsers.

[1] https://drafts.csswg.org/selectors-4/#the-blank-pseudo