remove all group from feeds before deleting
Main purpose is to prevent orphaned groups.

It's not possible to use the ON DELETE CASCADE trigger here, to remove
the group together with the last feed that is assigned to this group.

The ON DELETE CASCADE trigger will raise an foreign key violation error
in cases where the removed feed is not the last feed associated to a
group.

The purge_groups() call has been moved to the remove group functions,
since it's the only way to create an orphaned group.