bring back mark as unread/mark as read & bookmark links
The links were silently dropped with commit
4c0d975532ff3096a1f30d708d37b9651c132925 "Make clickable icons"
without any mention in the commit message.

Based on the commit title, I would assume this is an additional
functionality and not something that replaces existing stuff.

It's annoying to move the mouse across the screen (reading text from
the left to right) to get to the mark read icon for marking an article
as read.

All changes related to the removal of the links are reverted by keeping
the clickable icons.